requirement.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-versatileimagefield==2.2
djangorestframework==3.12.4
Pillow==8.3.2
python-magic==0.4.24
pytz==2021.3
sqlparse==0.4.2
